Refactored build-site script
