+++ website/lib/build-site.js
@@ -1,8 +1,7 @@
 const {join} = require("path");
-const {readdir, readFile, exists, remove, ensureDir, copy} = require("./helper/file");
-const dayjs = require("dayjs");
-const marked = require("marked");
+const {exists, remove, ensureDir, copy} = require("./helper/file");
 const {renderTemplateAndWriteFile, renderTemplate} = require("./helper/template");
+const loadPosts = require("./loadPosts")
 
 const devUrl = "";
 const prodUrl = "";
@@ -11,60 +10,19 @@ const homePageTitle = "HttPlaceholder -  Quickly stub away any HTTP service";
 const tutorialsPageTitle = "HttPlaceholder - tutorials";
 const downloadsPageTitle = "HttPlaceholder - downloads";
 
-const loadPosts = async () => {
-    const postsPath = join(__dirname, "../src/posts");
-    const posts = (await readdir(postsPath))
-        .filter(f => f.endsWith(".md"));
-    const result = [];
-    const titleRegex = /^title: (.*)$/m;
-    const dateRegex = /^date: (.*)$/m;
-    const descriptionRegex = /^description: (.*)$/m;
-    for (let post of posts) {
-        const postPath = join(postsPath, post);
-        const fileContents = await readFile(postPath);
-        const titleMatches = fileContents.match(titleRegex);
-        if (!titleMatches) {
-            return Promise.reject(`No title set for post ${post}`);
-        }
-
-        const title = titleMatches[1];
-        const dateMatches = fileContents.match(dateRegex);
-        if (!dateMatches) {
-            return Promise.reject(`No date set for post ${post}`);
-        }
-
-        const rawDate = dateMatches[1];
-        const descriptionMatches = fileContents.match(descriptionRegex);
-        if (!descriptionMatches) {
-            return Promise.reject(`No description set for post ${post}`);
-        }
-
-        const description = descriptionMatches[1];
-
-        const date = dayjs(rawDate);
-        const cleanedPost = fileContents.split("---")[2].trim();
-        result.push({
-            postKey: post.replace(".md", ""),
-            title,
-            date,
-            description,
-            formattedDate: date.format("YYYY-MM-DD HH:mm"),
-            formattedDateShort: date.format("MMM DD 'YY"),
-            formattedDateRss: date.format("ddd, DD MMM YYYY"),
-            cleanedPost,
-            markdown: marked.marked(cleanedPost)
-        })
+const renderPage = async (distPath, rootUrl, htmlFileName, htmlDistFileName, pageTitle, extraProperties) => {
+    const htmlDestinationPath = join(distPath, htmlDistFileName);
+    let properties = {
+        pageTitle: pageTitle,
+        rootUrl
+    };
+    if (extraProperties) {
+        properties = Object.assign(properties, extraProperties)
     }
-
-    result.sort((a, b) => {
-        const unixA = a.date.unix();
-        const unixB = b.date.unix();
-        if (unixA > unixB) return -1;
-        if (unixA < unixB) return 1;
-        return 0;
-    });
-    return result;
-};
+    
+    properties.body = await renderTemplate(htmlFileName, properties);
+    await renderTemplateAndWriteFile("template.html", htmlDestinationPath, properties);
+}
 
 (async () => {
     try {
@@ -91,53 +49,18 @@ const loadPosts = async () => {
         for (const post of posts) {
             const postFolderPath = join(postsPath, post.postKey);
             await ensureDir(postFolderPath);
-            const postPath = join(postFolderPath, "index.html");
-            const renderedBlogPost = await renderTemplate("blog.html", {
+            
+            await renderPage(postFolderPath, rootUrl, "blog.html", "index.html", `${postPrefix} - ${post.title}`, {
                 title: post.title,
                 contents: post.markdown,
-                date: post.formattedDate,
-                rootUrl
-            });
-            await renderTemplateAndWriteFile("template.html", postPath, {
-                pageTitle: `${postPrefix} - ${post.title}`,
-                body: renderedBlogPost,
-                rootUrl
+                date: post.formattedDate
             });
         }
 
-        // Load and write home page.
-        const indexPath = join(distPath, "index.html");
-        const renderedHomePage = await renderTemplate("index.html", {
-            rootUrl
-        });
-        await renderTemplateAndWriteFile("template.html", indexPath, {
-            pageTitle: homePageTitle,
-            body: renderedHomePage,
-            rootUrl
-        });
-
-        // Load and write downloads page.
-        const downloadsPath = join(distPath, "download.html");
-        const renderedDownloadsPage = await renderTemplate("download.html", {
-            rootUrl
-        });
-        await renderTemplateAndWriteFile("template.html", downloadsPath, {
-            pageTitle: downloadsPageTitle,
-            body: renderedDownloadsPage,
-            rootUrl
-        });
-
-        // Load and write posts page.
-        const postsPagePath = join(postsPath, "index.html");
-        const renderedPostsPage = await renderTemplate("posts.html", {
-            posts,
-            rootUrl
-        });
-        await renderTemplateAndWriteFile("template.html", postsPagePath, {
-            pageTitle: tutorialsPageTitle,
-            body: renderedPostsPage,
-            rootUrl
-        });
+        // Load and write pages.
+        await renderPage(distPath, rootUrl, "index.html", "index.html", homePageTitle);
+        await renderPage(distPath, rootUrl, "download.html", "download.html", downloadsPageTitle);
+        await renderPage(postsPath, rootUrl, "posts.html", "index.html", tutorialsPageTitle, {posts});
     } catch (e) {
         console.error(e);
         process.exit(1);
