--- website/src/posts/iforgot-review-your-working-days.md
@@ -1,15 +1,0 @@
----
-title: IForgot - Review Your Working Days
-date: 2015-06-03T20:17:15+02:00
-description: I've created a little command line tool which captures a screenshot every once in a while (you can set the interval in minutes) and saves it to a specific location on your PC. I save them to my Dropbox. I also copied a link to the tool in my startup folder, so everytime my work laptop starts, the program opens. Now, when I need to go back in time; no problem! I've got a folder filled with screenshots.
----
-
-If you have a job and have to sit behind the PC a lot, you probably know the problem of not remembering what you've done earlier today (or even earlier this week). I also have the exact problem; for example, sometimes when I have to fill in the monthly hour form. Also handy for those daily standups.
-
-I've created a little command line tool which captures a screenshot every once in a while (you can set the interval in minutes) and saves it to a specific location on your PC. I save them to my Dropbox. I also copied a link to the tool in my startup folder, so everytime my work laptop starts, the program opens. Now, when I need to go back in time; no problem! I've got a folder filled with screenshots.
-
-I hope it might help out someone else (I know of tools like <http://timesnapper.com>, but they are paid, and I don't need that many functions).
-
-**[Download IForgot](https://github.com/dukeofharen/iforgot)**
-
-![iforgot](/static/img/iforgot.png)
