--- website/src/posts/windows-tools.md
@@ -1,124 +1,0 @@
----
-title: Windows Tools
-date: 2018-07-06T18:26:00+02:00
-description: Welcome to my ultimate Windows tools list. This is the list with tools I always install when I install a fresh copy of Windows. This list is constantly updated. If you have any good suggestions, I'm open to it :) Just leave a reaction.
----
-
-Welcome to my ultimate Windows tools list. This is the list with tools I always install when I install a fresh copy of Windows. This list is constantly updated. If you have any good suggestions, I'm open to it :) Just leave a reaction.
-
-# Dev tools
-- Notepad++ (https://notepad-plus-plus.org/download/v7.html)
-- Sublime Text (https://www.sublimetext.com/3)
-- XAMPP (https://www.apachefriends.org/index.html)
-- MySQL (http://dev.mysql.com/downloads/mysql/)
-- SQLYog (https://github.com/webyog/sqlyog-community/wiki/Downloads)
-- Node.js (https://nodejs.org/en/)
-- Oracle VirtualBox (https://www.virtualbox.org/wiki/Downloads)
-- dotPeek (https://www.jetbrains.com/decompiler/)
-- Fiddler (https://www.telerik.com/download/fiddler)
-- LINQPad (https://www.linqpad.net/Download.aspx)
-- Putty (http://www.chiark.greenend.org.uk/~sgtatham/putty/download.html)
-- Cmder (http://cmder.net/)
-- Remote Desktop Connection Manager (https://www.microsoft.com/en-us/download/details.aspx?id=44989)
-- Filezilla (https://filezilla-project.org/download.php?show_all=1)
-- Visual Studio Community (https://www.visualstudio.com/en-us/products/visual-studio-community-vs.aspx)
-- Ghostdoc (http://submain.com/download/ghostdoc/)
-- CodeMaid (http://www.codemaid.net/)
-- StyleCop (https://stylecop.codeplex.com/)
-- Roslynator
-- Visual Studio Code (https://code.visualstudio.com)
-- Instant Markdown
-- NSIS
-- PHP, Perl, ASM, Bash code format
-- C#
-- PowerShell
-- XML tools
-- Microsoft SQL Server (http://www.microsoft.com/nl-nl/server-cloud/products/sql-server/)
-- SQL Server Management Studio (https://docs.microsoft.com/en-us/sql/ssms/download-sql-server-management-studio-ssms?view=sql-server-2017)
-- Python (https://www.python.org/downloads/)
-- JRE (http://www.oracle.com/technetwork/java/javase/downloads/jre8-downloads-2133155.html)
-- Android Studio (https://developer.android.com/studio/index.html)
-- SoapUI (https://www.soapui.org/downloads/soapui.html)
-- WinSCP (https://winscp.net/eng/docs/lang:nl)
-- Andy (http://www.andyroid.net/)
-- One-Click SQL Restore (https://sqlbackupandftp.com/restore/)
-- DotLocal (https://github.com/dukeofharen/dotlocal)
-- Postman (https://www.getpostman.com/apps)
-- NSIS (http://nsis.sourceforge.net/Main_Page)
-- AutoHotKey (https://www.autohotkey.com/)
-- dotnet-script (https://github.com/filipw/dotnet-script)
-- SQLiteStudio (http://sqlitestudio.pl/)
-- Zeal (https://zealdocs.org/)
-- Database Browser (https://www.etl-tools.com/database-browser/overview.html)
-- FakeSMTP (http://nilhcem.com/FakeSMTP/index.html)
-- Notepad2 (http://www.flos-freeware.ch/notepad2.html)
-- ILSpy (https://github.com/icsharpcode/ILSpy)
-- SmartGit (https://www.syntevo.com/smartgit/download/)
-- Linux sub system for Windows (https://docs.microsoft.com/en-us/windows/wsl/install-win10)
-- HttPlaceholder (https://github.com/dukeofharen/httplaceholder)
-
-# Office tools
-- Mozilla Thunderbird (https://www.mozilla.org/nl/thunderbird/)
-- LibreOffice (https://www.libreoffice.org/)
-- Skype (https://www.skype.com/nl/download-skype/skype-for-computer/)
-- KeePass (http://keepass.info/download.html)
-- Sumatra PDF Reader (http://www.sumatrapdfreader.org/download-free-pdf-viewer.html)
-- Microsoft Office
-
-# Power tools
-- Recuva (https://www.piriform.com/recuva/download)
-- CCleaner (https://www.piriform.com/ccleaner/download)
-- WinDirStat (https://windirstat.net/download.html)
-- NimbleText (http://nimbletext.com/)
-- Total Commander (https://www.ghisler.com/download.htm)
-- Chocolatey (https://chocolatey.org/install)
-- WinHTTrack Website Copier (http://www.httrack.com/page/2/en/index.html)
-- TeamViewer (https://www.teamviewer.com/en/download/windows/)
-- ShutUp10 (https://www.oo-software.com/en/shutup10)
-- IForgot (https://github.com/dukeofharen/iforgot)
-- Unlocker (http://www.freenew.net/windows/iobit-unlocker-1-0/68837.htm)
-- QuickStart2 (https://github.com/dukeofharen/quickstart2)
-- AOMEI Partition Assistent (http://www.disk-partition.com/download-home.html)
-- FreeFileSync (http://www.freefilesync.org/download.php)
-- RAMMap (https://technet.microsoft.com/en-us/sysinternals/rammap.aspx)
-- Clipdiary (http://clipdiary.com/)
-- lister.exe (http://www.ghisler.com/lister/)
-- WinMerge (http://winmerge.org/downloads/?lang=en)
-- F.lux (https://justgetflux.com/)
-- A Ruler For Windows (http://www.arulerforwindows.com/)
-- WinRAR (http://www.rarlab.com/download.htm)
-- 7-Zip (http://www.7-zip.org/download.html)
-- MyUninstaller (http://www.nirsoft.net/utils/myuninst.html)
-- Process Explorer (https://technet.microsoft.com/en-us/sysinternals/processexplorer.aspx)
-- TeraCopy (http://www.codesector.com/teracopy)
-- Console Calculator (http://www.zoesoft.com/console-calculator/)
-- Everything (https://www.voidtools.com/)
-- Explorer++ (https://github.com/derceg/explorerplusplus)
-
-# Multimedia tools
-- Foobar2000 (http://www.foobar2000.org/download)
-- The Gimp (https://www.gimp.org/downloads/)
-- Paint.NET (http://www.getpaint.net/download.html)
-- ColorPic (http://www.iconico.com/download.aspx?app=ColorPic)
-- VLC player (http://www.videolan.org/vlc/)
-- IrfanView (http://www.irfanview.com/)
-- Greenshot (http://getgreenshot.org/downloads/)
-- LICEcap (http://www.cockos.com/licecap/)
-- Movie Maker (http://windows.microsoft.com/nl-nl/windows/movie-maker)
-- HandBrake video converter (https://handbrake.fr/)
-- Inkscape (https://inkscape.org/en/download/windows/)
-- Spotify
-
-# Browsers
-- Firefox (https://www.mozilla.org/en-US/firefox/new/)
-- Privacy Badger
-- uBlock Origin
-- HTTPS Everywhere
-- Google Chrome (https://www.google.nl/chrome/browser/desktop/)
-
-# Virus scanners / safety tools
-- AdwCleaner (https://toolslib.net/downloads/viewdownload/1-adwcleaner/)
-
-# Other tools
-- uTorrent (http://www.utorrent.com/intl/nl/)
-- WhatsApp Desktop (https://www.whatsapp.com/download/)
