+++ website/src/templates/posts.html
@@ -1,6 +1,4 @@
 <div class="portfolio">
-    <h1><a href="{{{rootUrl}}}">Ducode.org</a></h1>
-    <h2><a href="{{{rootUrl}}}/feed.xml">RSS feed</a></h2>
     {{#posts}}
     <h2><a href="{{{rootUrl}}}/posts/{{postKey}}">{{title}}</a> ({{formattedDateShort}})</h2>
     {{/posts}}
