+++ website/src/templates/template.html
@@ -23,7 +23,7 @@
         <li><a href="{{{rootUrl}}}" class="secondary">Home</a></li>
         <li><a href="{{{rootUrl}}}/download.html" class="secondary">Download</a></li>
         <li><a href="{{{rootUrl}}}/docs" class="secondary">Documentation</a></li>
-        <li><a href="{{{rootUrl}}}/tutorials.html" class="secondary">Tutorials</a></li>
+<!--        <li><a href="{{{rootUrl}}}/tutorials.html" class="secondary">Tutorials</a></li>--> <!-- TODO -->
         <li><a href="https://github.com/dukeofharen/httplaceholder/discussions" class="secondary">Forum</a></li>
         <li><a class="secondary" id="theme-switcher">
             <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="currentColor"
